:recycle: / #122 - ModalFooter의 버튼을 외부로 추출
